Fix up support for 12-bit mono in Touptek/Altair/Mallincam interfaces. Actually it's 12-bit data in 16-bit words.  Issue #45

+++ liboacam/Altairstate.h
@@ -42,7 +42,7 @@ typedef struct Altaircam_STATE {
   HToupCam		handle;
   // video mode settings
   int			maxBytesPerPixel;
-  int			currentBytesPerPixel;
+  float			currentBytesPerPixel;
   int			currentVideoFormat;
   // buffering for image transfers
   struct Altairbuffer*	buffers;
